Clear cargo clippy -D warnings across workspace (test lints + lib lints)
Automatic clippy --fix applied the mechanical test-code lints (useless vec!,
needless_borrow, double parens, etc.) across kernel/coin/settlement/atomic/node
test files, then manual fixes for what --fix leaves behind:

- x3-kernel: property_tests closure must-use Results bound via let _; removed
  unused MockDispatcher + collapsed identical if-blocks in mock; de-scaffolded
  unused alias and dead svm/evm_bytes in packet_integration_tests; replaced an
  always-true timestamp>=0 assertion with a real !=0 check.
- x3-coin: removed double parens; factored TeamVestingView tuple type alias.
- atomic-trade-engine: removed dead cross_vm_leg helper.
- x3-settlement-engine: unwrapped a redundant nested "mod tests"; tightened
  expired_locks type inference; kept the invariants multi-arg facade and the
  intrinsic-args dispatchable with documented clippy allows.
- runtime: dropped the deprecated RuntimeEvent assoc/cfg duplication, prefer
  iterators over index loops, derived Default for FreezeState, removed an
  identity op, and kept SDK-required runtime_version string/cfg idioms and the
  inline FraudProofV1 call payload with documented allows pending a scheduled
  runtime version/migration.
- x3-foundry-core unwrap and x3-kernel store Default impls landed earlier.

+++ node/benches/rpc_dex_latency.rs
@@ -37,7 +37,7 @@ impl MockDexRpcClient {
 
     async fn estimate_swap(&self, request: serde_json::Value) -> Result<serde_json::Value, String> {
         // Simulate processing time: 5-15ms typical
-        let jitter = (self.request_count.load(Ordering::Relaxed) % 10) as u64;
+        let jitter = (self.request_count.load(Ordering::Relaxed) % 10);
         tokio::time::sleep(Duration::from_millis(self.latency_ms_base + jitter)).await;
 
         self.request_count.fetch_add(1, Ordering::Relaxed);
@@ -60,7 +60,7 @@ impl MockDexRpcClient {
 
     async fn execute_swap(&self, request: serde_json::Value) -> Result<serde_json::Value, String> {
         // Execute is slower: 20-50ms typical (includes settlement intent creation)
-        let jitter = (self.request_count.load(Ordering::Relaxed) % 30) as u64;
+        let jitter = (self.request_count.load(Ordering::Relaxed) % 30);
         tokio::time::sleep(Duration::from_millis(self.latency_ms_base * 2 + jitter)).await;
 
         self.request_count.fetch_add(1, Ordering::Relaxed);
